Move client context factory to new file
